Lift FixtureList filtering to Fixtures container
